Ben zeta

+++ client/src/components/Navbar.tsx
@@ -1,26 +1,48 @@
 import { useNavigate } from "react-router";
 import { baseUrl } from "../constants/baseUrl";
 import axios from "axios";
+import { useEffect, useState } from "react";
 import { useSound } from "../context/SoundContext";
 
 export default function Navbar() {
   const navigate = useNavigate();
+  const [role, setRole] = useState<string>("");
   const { isPlaying, toggleAudio } = useSound();
 
   async function handleLogout() {
     try {
+      if (role === "Staff") {
+        await axios.delete(`${baseUrl}/delete-user`, {
+          headers: {
+            Authorization: `Bearer ${localStorage.access_token}`,
+          },
+        });
+      }
+
       localStorage.clear();
       navigate("/login");
-      await axios.delete(`${baseUrl}/delete-user`, {
+    } catch (error) {
+      console.log(error);
+    }
+  }
+
+  async function getUserDetail() {
+    try {
+      const { data } = await axios.get(`${baseUrl}/user_detail`, {
         headers: {
           Authorization: `Bearer ${localStorage.access_token}`,
         },
       });
 
+      setRole(data.role);
     } catch (error) {
       console.log(error);
     }
   }
+
+  useEffect(() => {
+    getUserDetail();
+  }, []);
 
   return (
     <nav className="flex justify-center h-full items-center bg-hidden  w-full p-5 gap-x-20 bg-purple-800">
@@ -30,9 +52,7 @@ export default function Navbar() {
           alt="logout"
           className="w-10 animate-bounceUp"
         />
-        <span className="font-bold text-sm text-white animate-bounceUp">
-          Logout
-        </span>
+        <span className="font-bold text-sm text-white animate-bounceUp">Logout</span>
       </button>
 
       <a href="/">
@@ -41,9 +61,7 @@ export default function Navbar() {
           alt="home"
           className="w-10 animate-bounceUp "
         />
-        <span className="font-bold text-sm text-white animate-bounceUp">
-          Home
-        </span>
+        <span className="font-bold text-sm text-white animate-bounceUp">Home</span>
       </a>
 
       <button onClick={toggleAudio}>
@@ -52,9 +70,7 @@ export default function Navbar() {
           alt="sound"
           className="w-10 animate-bounceUp "
         />
-        <span className="font-bold text-sm text-white animate-bounceUp">
-          {isPlaying ? "Pause Audio" : "Play Audio"}
-        </span>
+        <span className="font-bold text-sm text-white animate-bounceUp">{isPlaying ? "Pause Audio" : "Play Audio"}</span>
       </button>
     </nav>
   );
